modify the control file years and create a new one
The control file of the multiyear cannot be used for squashed
simulations since year_begin and year_end are different. These two
parameters are used for running simulations and post
processing. For squashed simulations, year_end has to be the same as
year_begin. We create a new control file that is identical to the
multiyear control file with the above described property.